% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_paths_relative.R
\name{add_relative_paths}
\alias{add_relative_paths}
\title{Finds all CSV files with local paths}
\usage{
add_relative_paths()
}
\value{
list of CSV files to be filtered
}
\description{
Looks for all CSV files in project directory that have columns with a path
that is local (aka not relative), then adds columns and saves out as a new file.
}
