% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complaints.R
\name{download_all_complaints_prints}
\alias{download_all_complaints_prints}
\title{Downloads Complaints as HTML files}
\usage{
download_all_complaints_prints(path)
}
\arguments{
\item{path}{destination for files to be saved in.}
}
\value{

}
\description{
Downloads the "Print" version of all complaints and saves as an html file
}
