% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{download_all_incident_prints}
\alias{download_all_incident_prints}
\title{Downloads Behavioral Incidents as HTML files}
\usage{
download_all_incident_prints(path)
}
\arguments{
\item{path}{destination for files to be saved in.}
}
\description{
Downloads the "Print" version of all incidents and saves as an html file
}
