% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iforms.R
\name{download_iforms}
\alias{download_iforms}
\title{Downloads all iForms}
\usage{
download_iforms(
  path,
  start_date,
  end_date,
  dir_template = "\%Y-\%m-\%d \%H\%M\%S iForms Export",
  filename_template = "{ID}_{iForm}.html",
  sanitize_replacement = "-"
)
}
\arguments{
\item{path}{The root destination for files to be saved in.}

\item{start_date}{"From" date parameter. Either a date object or a character in Y-m-d format.}

\item{end_date}{"To" date parameter. Either a date object or a character in Y-m-d format.}

\item{dir_template}{String naming template for the directory all exports are
created in. This is passed through
\code{\link[base:format.Date]{format.Date}}. See \url{https://strftime.org}
for full list of date-time codes. A sensible default is set.}

\item{sanitize_replacement}{String passed through to
\code{\link[fs:path_sanitize]{fs::path_sanitize}} to replace control and
reserved characters in filenames. Defaults to "-".}
}
\value{
Returns list of directories and files created, for logging or message
purposes.
}
\description{
Provided with a path for downloading the iForms, \code{download_iforms} will
reate a timestamped directory for the export, download the "Print" version of
all iForms and save as a clean HTML file (no scripts, etc.) in directories
organized by Category and iForm name, save attachments in sub-directories,
and save a CSV master table listing all iForms, the number of attachments,
and their path (relative to the root path).
}
\details{
Downloads the "Print" version of all iForms and saves as an html file. Saves
CSV master table in CSV format in root. Files are organized within parent
directories by category and child directories by iForm name. Files are named
with the ID number and iForm name.
}
\section{Details This function will create an rvest session using stored}{
authentication parameters. If not set, it will prompt in interactive mode
to set those parameters. This function will overwrite any existing files
without asking or notifying; using the timestamped directory for export
should mitigate data loss risks but be aware. iForm names and Category
names in VAIRKKO can use any character; these names are passed through
\code{\link[fs:path_sanitize]{fs::path_sanitize}} to remove control
characters, reserved characters, and reserved or invalid periods and
spaces.
}

