% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/behavioral.R
\name{download_incident_files}
\alias{download_incident_files}
\title{Downloads incident files}
\usage{
download_incident_files(incidents, path)
}
\arguments{
\item{incidents}{Incidents tibble}

\item{path}{Path to save files to}
}
\value{
returns input invisibly
}
\description{
Walks through incident files list column and saves files to designated path.
}
